Use Nix to fetch ATOM feed
Fetching external resources during the actual build became problematic
after the last major Nix update (build purity was improved).

The sensible way of doing things in Nix seems to be to declare ALL
build dependencies prior to the actual build -- this includes
e.g. ATOM feeds to be aggregated.

I am probably not going to do the same for the individual repository
feeds as this feature is going to be replaced with a more general
_microblog_-like subpage.